Attempt to fix and normalize generation across platforms.
The primary change here is to stop using the sorted list.
The regen:xplatform does not sort, and just uses the order
provided in the .rb files. Sorting in the platform files
appears to have been added only to determine min/max values for
each constant set.

This change was needed because the sort was broken for
posix_fadvise constants on MacOS since fcntl required ok but none
of those constants exist. The resulting list was empty, got sorted,
and then sorted.first/last were nil for min/max. Fixing that led to
issues because all platforms will have some missing values
(.value = nil) so the sort by value must be constrained to min/max
purposes.

This will cause most platforms to regenerate with rather large
diffs, since the numeric ordering is now replaced with the order
from the gen/const files, but this should be a one-time hit.